Fix class property to support sphinx
Several changes were needed in order to get this working.
Most notably, rather than using a custom descriptor in all cases, use
`classmethod(property(...))` when operating on py3.9+. This allows
sphinx to handle this gracefully with its class property inspection
logic.

To implement, we switch `sdk_classproperty` to
`chainable_classmethod`, which only defines `classmethod` such that it
can be applied to other descriptors. On 3.9+,
`chainable_classmethod = classmethod`, and on lower pythons, it is our
custom descriptor class. It turns out that `classmethod` does not
create a data descriptor, so we drop `__set__` and `__delete__` from
our descriptor definition.

The main body of this, the `__get__` method of the descriptor, is
cribbed directly from python documentation. It has to check if it's
operating on a descriptor and use the descriptor as appropriate.
This turns out to be quite tricky to try to test, so the testsuite
only covers the "normal" cases and doesn't bother about any others.

Because sphinx runs in ReadTheDocs are configured to use py3.9, set
basepython=python3.9 for the docs tox env, to get better emulation of
the docs site build. If your default python is py3.8 or lower, the
docs would render differently based on the behavior of this
platform-supporting code. On 3.9, `classmethod` gets used, and the
sphinx inspection of descriptors works as desired to document a class
property.

Additionally, use `classmethod` as the decorator when operating under
`typing.TYPE_CHECKING`, i.e. a mypy run. This needs to hold even if
`mypy` is installed and invoked under python3.8 or lower.