Refactor classproperty to satisfy mypy and test
Rather than using classmethod(property(...)), which is not properly
understood by mypy, split the sphinx-build treatment of classproperty
from both runtime and mypy evaluation. Test the result to ensure the
type identified by reveal_type is accurate.

1. Under sphinx runs, a utility method `in_sphinx_build()` will return
   True; use this to dispatch on compatibility code
2. If `in_sphinx_build()` is true,
   `classproperty = classmethod(property(...))`
3. If `in_sphinx_build()` is false, `classproperty` applies the custom
   descriptor, `_classproperty` -- this applies both to runtime and
   mypy evaluation

In removing any attempt to support sphinx in the descriptor, it
becomes possible to make it simpler and more direct. As a generic
descriptor, mypy is able to correctly infer the type of the
resulting attribute from the return type of the decorated
function.

The result
- passes behavioral tests
- produces the right types in mypy
- generates class property docs in sphinx

Where this gets *really* interesting is in testing. We don't currently
have any way of writing unit tests which check the types we're
producing -- a gap in our current coverage.
mypy can't be used as a library, except via a crude "api" module which
intentionally provides the same basic interface as the CLI.

What this means is that in order to check the results of
`reveal_type()`, there's no way to avoid passing through strings. But
that doesn't mean that checking the types is impossible. Rather, we
can wrap any bit of code we want to check in a tiny file, dump that in
a tmp dir, and run `mypy.api.run` on the result. Parsing the output,
we can get any reveal_type information from the `note: ...` output.
Wrap all of this up in a test fixture, and we can test that
`reveal_type(BaseClient.resource_server)` is `Union[builtins.str,
None]`. In some ways, the interface is reminiscent of `timeit`, with
eval'ed python passed via strings.