Fix preservation of type info via has_paginator
`has_paginator` was meant to preserve type information, but it was not
doing so properly.

Because the inner decorator was correctly annotated but the outer
layer was not, all functions were being converted to generic Callable
types when decorated with this.

Additionally, although some cases for preservation of type information
in PaginatorTable are difficult (see GitHub #490), the simple case of
runtime typing for paginated methods can be handled easily by using
functools.wraps on the generated paginated methods.

It's possible in theory that this is inaccurate in some boundary cases,
but it's definitely more accurate than the current type information of
a generic callable accepting any arguments.

In order to test this change, a new mypy-testing utility is
introduced, `run_reveal_type`, specifically for testing typeA
annotations. It runs mypy's reveal_type() on a string input and
reports back the result.