Set mypy disallow_any_generics=true
This setting is in the same category as `no-implicit-optional'. When a
generic is specified in this mode, it _must_ have type arguments, even
if the type argument is "Any".

This avoids overly permissive and sometimes incorrect types in various
situations, e.g. when a decorator is typed as `Callable -> Callable`