Initial work on typed_paginator
This adds a first draft of 'typed_paginator' as a client method which
takes other client methods as inputs.

Unfortunately, because mypy v0.910 does not yet implement support for
ParamSpec, this code is not yet usable. `mypy` will error on ParamSpec
usage, indicating that it believes that the parameter specification
for a Callable must be a list or ellipsis. The implementation appears
correct (with the caveat that it uses `assert`s and is minimally
tested and documented) using mypy master, v0.920, but this means it is
not yet ready for general use.