Make Paginator a Generic type over PageT
This makes Paginator a generic over PageT, a type var with the
type bound of GlobusHTTPResponse. This means that hte iterable
elements of a paginator are still expected to be response objects, and
they will still type-check as having the necessary methods and
attributes of a response. However, this also makes it possible for a
paginator to be constructed with PageT bound to a specific value, e.g.
`IterableTransferResponse`.

This would mean that the pages of the paginator have a more specific
type, and can be expected to respond to those methods.

Under the current implementation of pagination, a lot of type
information is being lost. However, if a more type-specific pagination
interface is added (PR #494), then the type-preserving paginator will
be able to indicate the page type which it is producing.

Because `mypy --strict` disallows implicit `Any` on generics, several
locations need to be updated with an explicit type variable for the
paginators, usually Any.
One minor change this raised is that it allowed us to turn off a
`type: ignore` in the paginator table. The result, however, is that we
must add `**kwargs: Any` to the base paginator class (because
subclasses may implement additional keyword args) in order for that
method to type-check.