Expose `headers` from response objects (#497)
* Expose `headers` from response objects

`GlobusHTTPResponse.headers` is just a pointer at the
`requests.Response.headers` mapping.

To enable this, an internal-only property, `_raw_response` is added
which traverses wrapped response objects until it reaches the original
response object. We can then use that property to implement `headers`
as a property which returns `self._raw_response.headers`. This pattern
is also applicable to some existing data -- `http_status` and
`content_type` -- so these are changed to be uniform.

This means that `http_status` and `content_type` are now read-only,
which is a minor interface change. It is therefore noted explicitly in
the changelog.

* Add test for HTTP header unpacking

This is a behavior which is being added to allow us to fetch
headers from a (potentially nested) response object. Add a test
explicitly for this behavior.