Avoid building 'Paginator[Any]'
As we discovered, if a generic type sets a bound, the bound is
properly enforced for subtypes. However, if a subtype declares itself
with `Any` for the type parameter, then the bound is no longer
enforced.

For example, we can define a generic class which uses a type var to
bound its inputs, and then violate that bound by declaring an instance
with the type parameter Any:

    X = TypeVar("X", bound=str)

    class A(Generic[X]):
        def foo(self, x: X) -> X:
            return x

    b: A[Any] = A()
    b.foo(1)        # <--- type checks!

To avoid this situation, we need to ensure that `Paginator[Any]` is
not used in various locations. Because `PageT` is a type variable, it
needs to be scoped appropriately for its usages.