Initial pylint config + resolve some warnings
Setup a pylint config, set to ignore various formatting related
messages, which otherwise conflict with black.
Given that `*args, **kwargs` are often used as passthroughs to accept
and discard arguments intentionally, these are ignored when unused.
The docparams plugin was tested, but produces many warnings right now.
Disable for the time being.

Address some of the warnings:
- unused `**kw` (update to `**kwargs`)
- redefined `cast` in `env_vars` (also resolve some typing complexity
  with better overloads)
- `raise` rather than `raise ... from ...`
- one undocumented `:raises:` from before disabling docparams
- incorrect unsubscriptable-object is ignored on type subscripting