Always use UTF-8 when doing file IO
Whenever `open()` is called, pass `encoding="utf-8"`. This is the
de-facto standard for text encoding and is unlikely to cause users any
issues.

The most likely cases for file IO are ASCII-only. Although this is
technically minorly backwards incompatible, chances are it will have
no deleterious impact on users. Furthermore, there are changes in
behavior for python which are happening around bare `open()` calls:
  https://www.python.org/dev/peps/pep-0597

As PEP597 notes, the previous default behavior could be requested
explicitly with

    open(..., encoding=locale.getpreferredencoding(False))

However, this assumes that we want to make the locale-specified
encoding the default for the SDK. That is not the case -- it would be
better to always use UTF-8 and be explicit where we care.

The only case today in which this may be incorrect is GridFTP gridmap
for GCP. If it turns out that this is encoded in latin-1 or the user's
preferred encoding, we will adjust the encoding used to read the file
to match the GridFTP behavior. This is assumed to not be a problem
until proven otherwise.