Fix where IdentityMap checks lookups against cache
Whenever a lookup is happening, we should check the cache to see if it
was populated (for shared cache use-cases). However, this hsouldn't
only happen on `__getitem__` when a new key is being looked up.

Instead, whenever building a batch to send to the API as a request,
the elements in the batch should all be checked against the cache.