Add pagination to `SearchClient.post_search`
Enable paginated searches via this method.

Pagination is different in this API from any existing APIs because
`limit` and `offset` are sent as body parameters. As a result, some
changes need to be made. The client method must support passing these
and setting them in the passed data (as overrides).

In the case that these are passed, the body therefore needs to be
(shallow) copied before being updated. Tests ensure that this is done
correctly, and that paging will therefore not be side-effecting.