Add SearchClient.scroll and SearchScrollQuery
A new method of `SearchClient`, `scroll`, allows access to the
scrolling API. `SearchScrollQuery` is a helper type for easily
building query documents. Scroll queries are similar to typical
searches, but do not support all of the same fields, including
boosts and facets. As a result, to support this alternate type, it and
SearchQuery now share an internal base class which provides several of
their shared methods.

A new test exercises `paginated.scroll`, confirming that it works as
expected with both dict data and `SearchScrollQuery`. Like paginated
`post_search`, the method is written carefully to not be
side-effecting on the data it is passed, and the test confirms this.