Fix documentation of search helper objects
show-inheritance in autodoc behaves badly by default on these classes
because they inherit from UserDict, and it therefore shows all of
their dict-like methods.
In order to have clear, but non-repetitious, documentation, the
SearchQueryBase class is now documented, with a note that it is only
included in docs to show the methods it provides.

None of the methods of the search helpers were being picked up under
autodoc because they had no docstrings. To rectify, add docstrings to
all methods.