First draft of an internal _testing package
Use module names to namespace and organize the fixture data, and
define a 'ResponseSet' class to wrap the responses and provide
naming, access, etc.

Various methods for loading fixtures return a RegisteredResponse
object, which can have a metadata dict and several pieces of
information about the mocking which was done.

To demonstrate the new tooling, set it up and begin using it for some
of the simpler Auth tests.