Set a floor for versions of typing_extensions
`typing_extensions>=4` is required (where typing_extensions is
required).

This is tested in tox both with `mypy-mindeps` and with
`py36-mindeps`. The mypy run is only done in GH Actions to reduce the
time spent on this when running `make lint`, as it is unlikely to
fail.

The `typing` repo has a discussion about capping or not capping the
version of `typing_extensions`, but it is not well-resolved. For now,
do not apply a cap to the version of `typing_extensions`. We can
revisit this if the lack of a cap ever becomes an issue.