Use HTTP "reason" as failover in exc args
When marshalling arguments to pass to `Exception` for an API error,
the `message` field is included last. However, in the event of an
empty body, there's no use for that field.

HTTP provides for a "reason phrase", e.g. the "Not Found" in
"404 Not Found".

With this change, the reason will be used as a failover for the empty
message case. No public members of the exception classes are changed
-- this is done only to provide slightly better traces when errors are
sent back by API Gateway, WAF, and other proxies and devices which
might send a useful "reason" but discard or not generate a useful
body.