Update to responses==0.17.0 and fix
This is the final `responses` version with py3.6 support, so we can't
upgrade past this until we're ready to begin dropping 3.6.
Updating means that test tools need to set `match_querystring=None` to
avoid a deprecation warning. This should ensure that any consumers of
`_testing` are not impacted with deprecation warnings (so long as they
keep to this same version of `responses`.