Add support for 'include' as a param to get_group
The primary change here is to add support for
`get_group(..., include="memberships")` and related usages. We'll need
this for `globus group members list` which will pass this parameter.

However, there are two "adjacent" changes being included here for
expediency. One is to move some of the test fixture data into
`_testing` -- only for tests which are being modified now. The other
is to ensure that the docstrings on all GroupsClient methods include
param docs.