Ensure that IdentityMap marks keys resolved
The unresolved lists in the IdentityMap should be updated whenever a
key is looked up. Otherwise, the only check which avoids duplicate
lookups is a check for values already in the cache. This could fail in
some cases, e.g. keys which do not actually resolve to a value
(nonexistent usernames).