Expose more details of HTTP requests
New options:

- allow_redirects: bool on `request` methods
- stream: bool on `request` methods

These are underlying `requests` options which are re-exposed via the
SDK methods for raw request sending. Because they are "advanced", they
are not also re-exposed on the `get()`, `post()`, etc helper methods.

New properties:

- http_reason on responses and errors
- headers on errors

This lets the caller get more detail about the HTTP interaction in the
success and error cases.

Tests exercise the new functionality, with the caveat that
`allow_redirects=False` doesn't really do anything when running under
`responses`. If we find a better test, there's a comment which will
help us know that we can remove the old test.

For the tests, there are some status->reason mappings that "do the
normal thing", e.g. `200 -> OK`, but these aren't baked into the SDK
anywhere or even into `_testing`.