Add MutableScope type for manipulating scopes
In order to define optional scopes and scoes with dependencies, define
a new type, MutableScope, from the scopes module. Anywhere that scope
strings are accepted, MutableScopes are now accepted too.

MutableScope.collection2str is a static method attached to the class
for converting strings, iterables of strings, and mixed iterables of
strings and MutableScopes into the desired (space-separated) string.

Additionally, `ScopeBuilder.make_mutable` is added as a method for
ScopeBuilders to create MutableScope objects from their existing scope
attributes.

Narrative docs are added and MutableScope is added to the `scopes`
module's reference documentation.