Improve handling of 'include' for storage gateways
Improve the handling of the 'include' parameter in GCSClient, and add
tests to cover the various usages. Effectively, if you pass a string,
you should see that string pass through. If you pass an iterable,
you should see it joined with commas. And if you pass None, you should
not see the parameter.