Minor tweaks to test config
- add a `find` to `make clean` so that it scrubs `__pycache__` dirs as
  well
- set pytest testpaths=... for faster test collection (avoids scanning
  any other dirs for test modules)
- set a needed ignore for cryptography warnings for py3.6 builds

Regarding the last item, the constraints for handling the ignore are
tied up in the way that the `warnings` module is initialized. We
cannot give a very precise ignore, so use string-based matching for
now. If this turns out to be unstable, we can revisit and potentially
disable all warnings for the 3.6 builds via tox factor config.