Add tests for untested transport behaviors
There were two lines of the 'transport' subpackage which were untested
in coverage reporting. In each case, a couple of simple unit tests are
sufficient.

1. RetryCheckRunner behaviors

Current testing was not checking the case of a RetryCheckRunner
returning `False` because a check responded with `do_not_retry`. Check
this case, its converse (`do_retry`), and the fallthrough case in
explicit unit tests.

2. Authorization header handling in RequestsTransport

In a full request sending context, a RequestsTransport may use a
provided GlobusAuthorizer to set an Authorization header on a request
and update it after retries. Our functional tests cover several cases,
but one which is missing is the NullAuthorizer case (which should
clear Authorization even if it was explicitly provided). Rather than
adding more functional tests, this is isolated from its surrounding
context by a few unit tests which check the exact behavior of
`RequestsTransport._set_authz_header` to ensure it is what we expect.
Functional tests could be added in the future as well, but are not
included here.