Implement lazy import semantics using __getattr__
Using module-level getattr (py3.7+), implement a lazy importer for the
top-level SDK imports.
These are duplicated in a way which exposes names to mypy correctly.

This revealed several import cycles with authorizers and auth client
and response types. The cycles are broken here using TYPE_CHECKING
guards (since the types are the source of the cycles).

A new script at `src/globus_sdk/__generate_init.py` autogenerates the
__init__.py for the SDK. This allows the following

- specify top-level imports only once and autogenerate the duplicated
  content between the import table, the __all__ and the actual imports
- use the concrete import path on py3.6 to allow compatibility with
  that python version