Add support for more StorageGatewayPolicies docs
All of the remaining (current) types from gcs-cli have been added
here. Most are a matter of copy-paste-replace with minor adjustments
to docstrings.

One item in here merits more attention. `OneDriveStoragePolicies`
supported "EXPLICIT_NULL" in the gcs-cli. This type of value has been
used in globus-cli as well to indicate a value which should be set to
`null`/`None` instead of being omitted, but it has never had any
representation in the SDK. Now, `utils.missing` is a sentinel value
(with type `utils.MissingType` which can be used for this purpose. The
field in question defaults to `utils.missing` and is checked for
`None` explicitly. The PayloadWrapper helper methods recognize
`missing` as an alternative to `None` and omit it.

In the future, the SDK may want to support `missing` as the default
for all helper object fields, so that setting `x=None` explicitly
always has an effect. This could be used throughout the SDK, even
beyond helper objects. However, that change is breaking and would
require a major release.