Update fields to extract AuthAPIError messages
"message" does not appear to match the API -- it is never used.
"error" is a copy of "code" in some cases, but it is only used in the
top-level error wrapper (not the nested error doc we extract).
"title" *is* actively used as an alternative to "detail" and was not
being checked. Add this to the fields.

This can be refined further over time if we see issues with this
message parsing.

The change could technically be breaking for code which uses
`AuthAPIError.message` to drive logic. e.g.

    try: ...
    except AuthAPIError as e:
        if "foostr" in e.message: ...
        raise

However, we're improving behavior and staying consistent with the
spirit of semver. The above kind of logic presumably _should have
been_ written as

    try: ...
    except AuthAPIError as e:
        if "foostr" in e.raw_text: ...
        raise