Renames for generate-init script
- rename `__generate_init.py` to `_generate_init.py` absent any strong
  motivation for the double underscore
- switch `globus_sdk.types` to `globus_sdk._types` because of the
  impact of a module named `types`

In certain run contexts -- primarily, with a typical dev virtualenv
active -- `_generate_init.py` would fail. The cause was ultimately that
the importer would pick up `globus_sdk.types` in place of the stdlib
`types` module.

We have no need or desire to publish the `types` module from the SDK.
Rename it to `_types` in order to ensure that there is no risk of this
happening when the pre-commit hook is executing (e.g. `git commit`
with a virtualenv active).