Add ArrayResponse for GroupsClient.get_my_groups
ArrayResponse is a new response type similar to IterableResponse which
defines `__iter__` on the raw `data` itself. As such, it benefits from
the addition of top-level `__bool__` and `__len__` so that it can be
tested and inspected similarly to any other Sequence type. It also
benefits from being able to use non-string keys (ints and slices) in
__getitem__.

Adding ArrayResponse therefore folds in these other improvements:
- `__bool__` which can now be false. A very strict reading could treat
  this as backwards incompatible for niche cases
- `__len__` is now defined (also works on dict data)
- `__getitem__` now takes `str | int | slice`, not just `str`