Add a complete example with tokenstorage
This is a near-duplicate of the group listing example script, but it
stores the user's creds using tokenstorage in a JSON file in their
homedir.

Add the example, and also cross-link it from the tokenstorage docs.

Fix some minor errors in the existing tokenstorage example (wrong key
for "expires_at_seconds").