Improve testing subpackage with more responses support and inherited metadata (#584)
* Update to use the latest 'responses' on py3.7+

This is an update to the testing requirements to use different
versions of `responses` on different python versions. It *should*
allow use of the latest `responses` version with `globus_sdk._testing`
without warnings or errors.

As a related tweak, this change revealed that we still have one test
which was using `six` (installed via the `responses` dependency). It
is easily removed.

* RegisteredResponse objects inherit metadata

When registered as part of a ResponseSet, a RegisteredResponse will
- mark the ResponseSet as its `parent`
- setup `metadata` to refer to `parent.metadata` IFF `metadata` was
  not supplied on init

Adding a RegisteredResponse to a ResponseSet does not mark `parent`,
but the attribute can (of course) be written.

Update all tests to skip setting `metadata=...` on RegisteredResponse
objects where the value is inherited from ResponseSet.