Allow use of non-default requests_mock in _testing (#596)
* Allow use of non-default requests_mock in _testing

responses has an interface which allows you to customize the mocking
of `requests` with an object. The `_testing` package was all built
using the default mocks, which meant that it couldn't accept a
`RequestsMock` as an input and gracefully attach to it.
This adds `requests_mock=...` as an argument to all response loading
methods.

Also, extend the readme with
- doc on RequestsMock usage
- a short "index" of _testing contents

* Add some tests for non-default RequestsMock

This is part of _testing being tested, so it's hard to see a way of
doing it without getting tied to some of the methods we have
throughout the SDK. Choose a relatively simple one (get_group) as an
example.

Not all methods are tested here, just a couple of essentials.