Add __dir__ to the top-level SDK module (#603)
* Add __dir__ to the top-level SDK module

Explicit __dir__ is part of the PEP 562 spec which added module-level
__getattr__. We were missing this function, which had limited negative
impact other than the fact that `globus_sdk.Tr<TAB>` was not handled
nicely by an interactive interpreter.

* Refactor dir(globus_sdk) to include more names

Add all of __all__ as the default, and then include __all__, __file__,
and __path__. This is a judgement call in terms of which module
attributes are useful to have in `dir()` output. `__loader__` and
`__spec__`, for example, may be useful in theory for some users.
However, they are more likely to be confusing to the average user,
whereas `__path__` is a useful diagnostic when multiple virtualenvs
are in use.