Update TransferData and DeleteData to allow no `submission_id`, then fit this with TimerJob.from_transfer_data (#602)
* Improve TransferData and DeleteData null handling

Convert various argument usages to the null-removing setters and
ensure that `add_item` removes null arguments instead of passing them
to the API.

Allow transfer_client=None in order to support a usage in which
`submission_id` is omitted for both TransferData and DeleteData.

Convert `skip_activation_check` to an Optional with a default of
None so that it will be omitted by default.

* Update TimerJob to reject bad TransferData

If a submission_id or skip_activation_check value is present, then the
TransferData won't translate cleanly into a TimerJob, so it should be
rejected.

In functional tests, the use of a TransferClient is dropped from
TimerJob.from_transfer_data usage.

* Changelog for Transfer and Timer helper changes

* TransferData and DeleteData allow kwarg-only usage

In these helper objects, set `transfer_client`,
`source_endpoint`, `destination_endpoint`, and `endpoint` to optional
with defaults of `None`. They are now optional positional arguments,
meaning that they can be passed as named parameters, instead of
requiring a first positional argument of `None` for the
null-TransferClient usage.

As a result, the `source_endpoint`, `destination_endpoint`, and
`endpoint` parameters need to be explicitly checked at runtime to
ensure they are not passed as `None`. Although this slightly weakens
the capabilities of type checking, it allows for better runtime
semantics and usage in the good case.

The testsuite is updated to test these behaviors and ensure that bad
arguments are rejected. As the usage is now simpler, some testing
helpers are no longer useful and have been removed.

Two related but distinct changes are added here in support of Timer
usage:
- TimerJob now logs an info message on init (useful for
  troubleshooting)
- `recursive_symlinks` is now Optional with a default of `None`,
  meaning it will be omitted from TransferData by default

* submit_{transfer,delete} auto-set submission_id

TransferClient task submission methods automatically request and set
the submission_id if one is not seen in the data. New tests cover this
behavior in detail.