Add pagination back to FlowsClient.list_flows (#621)
* Add pagination back to FlowsClient.list_flows

The bulk of this change is a response data generator which produces
test data. Although it would be nice to put this set of responses into
the `_testing` data, it's not 100% clear *where* it should go. For
now, this changeset punts on that question and puts the test data into
the test module.

Now that Flows is supporting `marker`, the most barebones application
of the pagination decorator for marker pagination is sufficient.

* Remove use of fromisoformat

* Use a sys.version_info dispatch on fromisoformat

`pyupgrade` will automatically "upgrade" this codepath to the py3.7+
variant once we drop support for py3.6