Support python 3.11 (#628)
* Support python 3.11

In the changelog, this is called out as "official support" since the
code all worked on 3.11 prior to this change.

Places touched:
- CI testing job on linux (all pythons)
- CI testing on windows (latest python only)
- CI testing on macos (latest python only)
- readthedocs config
- pypi classifiers
- tox testenvs

CI builds should be examined to see what might fail in particular.
Local testing under pyenv+tox works, but it's always possible that
something will be different in the CI environment.

* Add a close() method to SQLiteAdapter

This is being done to fix the behavior of tests which are leaking
resources. The leaked connections can cause tests to fail because
there's still a handle on the `.db` file when the temp dir tree is
being removed.

The method makes it possible to close the connection through a public
interface. Rather than adding a context manager interface, for now the
testsuite operates by having a fixture which auto-closes SQLiteAdapter
instances during test teardown.