Convert GCS get_collection_list tests to new style (#635)
* Convert GCS get_collection_list tests to new style

This change moves the tests for this method into a dedicated module,
and moves the test data from the older `fixture_data/` dir to the
`globus_sdk._testing.data` subpackage.

The oddity here is that get_collection_list is being used to test
error parsing for a GCS 403. It's not immediately clear where this
should go, so it's bundled with these other tests for now. (Perhaps
this test is good but should be augmented by additional unit tests.)

* Update src/globus_sdk/_testing/data/globus_connect_server/get_collection_list.py

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>

* Separate get_collection_list parameter test cases

Per review, break a monolithic test into several component tests with
parametrization for their inputs where appropriate.

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>