Fix bug with str(OAuthDependentTokenResponse)
The stringifier inherited from OAuthTokenResponse assumes that
`self.get(...)` is safe, but that's only true when the data is a
`dict`, as it normally is.
OAuthDependentTokenResponse has `self.data: list`. Therefore, this
needs a check and special handling in the stringifier.

Add a test case which fails without this addition, then make the
`get()` call conditional on the type of the underlying data.

Arguably this is a "bad" division of responsibilities (base class
knows about its subclasses), but any alternative which solves this
generically is more verbose or touches on core codepaths like the
central `GlobusHTTPResponse.get` implementation.