Allow `get()` when responses contain lists
`GlobusHTTPResponse.get` checks if the data is `None` for special
behavior, but this check can easily be extended to the case in which
`data` is a list. This makes `get()` safer to use when data may be
a dict or list, as it won't try to user `list.get` and encounter an
AttributeError.

Add a changelog fragment to this effect.

The result is also that the changes to `OAuthTokenResponse` to handle
the `OauthDependentTokenResponse` are no longer necessary and are
reverted here.