Fix bug with str(OAuthDependentTokenResponse) (#644)
* Fix bug with str(OAuthDependentTokenResponse)

The stringifier inherited from OAuthTokenResponse assumes that
`self.get(...)` is safe, but that's only true when the data is a
`dict`, as it normally is.
OAuthDependentTokenResponse has `self.data: list`. Therefore, this
needs a check and special handling in the stringifier.

Add a test case which fails without this addition, then make the
`get()` call conditional on the type of the underlying data.

Arguably this is a "bad" division of responsibilities (base class
knows about its subclasses), but any alternative which solves this
generically is more verbose or touches on core codepaths like the
central `GlobusHTTPResponse.get` implementation.

* Allow `get()` when responses contain lists

`GlobusHTTPResponse.get` checks if the data is `None` for special
behavior, but this check can easily be extended to the case in which
`data` is a list. This makes `get()` safer to use when data may be
a dict or list, as it won't try to user `list.get` and encounter an
AttributeError.

Add a changelog fragment to this effect.

The result is also that the changes to `OAuthTokenResponse` to handle
the `OauthDependentTokenResponse` are no longer necessary and are
reverted here.