Add scope parsing to Scope object (#642)
* Add scope parsing to Scope object

The core of this change is that `MutableScope` has been renamed to
`Scope` and now has a pair of parsing functions as classmethods:
`Scope.parse` and `Scope.deserialize`. The first parses any valid
scope string to request, meaning it may be a space delimited list of
scopes, and the second parses any valid scope string with a single
top-level scope.

As an inverse of `deserialize()`, the stringifier has been renamed to
`serialize()`, and `__str__()` now just calls `serialize()`.

The parsing itself requires several new internal components. A
tokenizer handles converting strings into lists of tokens, e.g.
converting `foo[bar *baz]` into
`["foo", "[", "bar", "*", "baz", "]"]`.
The tokens are small objects pairing strings with a type, which makes
it simpler for the parser to handle various cases. The parser is
minimally stateful, has no recursion or backtracking, operates with
limited lookahead, and is generally meant to be the smallest and
simplest thing which can do the job that it does.

The parse result, a `Scope`, can now represent the true tree-like
structure of scopes. `Scope.dependencies` is a list of child nodes in
the tree, a list of  `Scope`s. This has implicitations for
`Scope.add_dependency` in that a `Scope` is now a valid input. As a
result of this introduction, the use of
`add_dependency("foo", optional=True)` is considered deprecated and
now emits a DeprecationWarning. Building dependencies in the current
version should be done either with (newly parseable) strings or with
`Scope` objects. i.e. The aforementioned usage can be replaced with
either `add_dependency("*foo")` or
`add_dependency(Scope("foo", optional=True))`.

Some related changes included here:
- `globus_sdk.scopes` is now a subpackage (it was getting a bit long)
- the internal type alias for ScopeCollectionType is now kept in
  `_types` for simpler universal access within the SDK
- many more unit tests are added for the scope parsing and
  serialization functionality
- documentation is updated to refer to these as "Scope objects" rather
  than "MutableScopes"
- various doc fixes, including missing scope data in the scopes docs
  and a bug in the listknownscopes custom directive

* Forbid special characters in Scope.__init__

This avoids having to consider parsing or otherwise handling
`Scope("*foo")`. Although this would ideally be an alias for
`Scope.deserialize("*foo")`, the interface gets messy if we allow
this.

* Add `Scope.__contains__` with tests

`Scope(...) in Scope(...)` is meant to indicate "permissions
containment", meaning that the permissions requested by the first
scope are covered by the permissions requested by the second scope.

That notion of containment leads to the following defined behaviors:
- the top-level scope strings must match
- either both have matching optional=t/false OR only the
  "contained" scope is optional
- each dependency needs to be contained in the "containing" scope's
  dependencies
- it doesn't make sense to ask if a non-Scope value is in a Scope

This definition feeds directly into an implementation and a large
suite of tests to check correctness.

* Allow optional=t/f on make_mutable

This allows better scope constructor behavior from
`ScopeBuilder.make_mutable`, as an optional scope can be constructed
in a single step.

* Refine `Scope.__contains__` and update doc

- Ensure `__contains__` is included in docs
- Rebrand the `__contains__` check as "permission coverage" throughout
  relevant doc
- Update the structure of the check to use simpler control-flow
  (a "reminder" comment should help make the for-else readable to
  those not familiar with the construct)

* Consolidate scope parsing tests and add more cases

Several cases of scopes rejected by Globus Auth but previously
accepted by this parser have been added. In particular:
`foo[bar]baz` is not allowed, nor is `foo [bar]`. Both of these
strings were previously allowed by the parser (but would re-serialize
correctly for Globus Auth consumption).

In order to support this, the lexer now also implements a "peek" value
for lookahead and checks it against various possible values. This
enables simpler checks than would otherwise be possible when iterating
over the input string.

* Add more test cases for scope parsing

* Make Scope "contains" a private helper method

We retain all of the tests, but the method is renamed to `_contains`,
removed from public docs, and commented as a "non-authoritative
convenience function". In particular, after discussion we do not
believe that it is valuable to try to build fully accurate
client-side scope comparison when this cannot be authoratative with
respect to a user's current consents in Auth.

For now, the method will exist and it will be available for our own
use-cases (e.g. in globus-cli), but not for general use.

Also, improve documentation of `add_dependency` to avoid (confusing)
chaining usage.

* Add scope parsing tests which only assert validity

These tests do not check anything *about* the parsed scopes. Only that
the strings which are being fed to `parse()` are valid.

* Test scope parsing on service formatted scopes