Prototype scope graph parsing
A scope graph properly represents the data structure being used when
Globus Auth evaluates scope requests.
This work handles the conversion of a ScopeTreeNode representation
into a ScopeGraphNode representation, converting a set of Trees to
Graphs. It also preserves the set of top-level scopes, as this is
relevant in terms of the return type seen by a requester.

For now, the Scope class is broken by this change. It is not clear
what a Scope object should be with respect to the graph which can be
extracted from parsing. Keeping them fully distinct is bad for the
case of loading a scope string (which was saved), then modifying and
requesting it.