Simplify graph representation of scopes
Rather than representing nodes as objects, just use the scope strings
as the nodes and tuples of scope strings as the edges.

The graphviz dot rendering of a scope graph is now implemented as the
`__str__` of a graph object, and the graph itself tracks the top-level
scopes.

Because graphviz dot data does not require that all nodes are declared
a-priori, the str() method renders the top level scopes as declared
nodes, which makes them slightly more visible to a reader.