Yank scope parsing, revert to MutableScope (#654)
* Yank scope parsing, revert to MutableScope

After discussion of the ways in which a parse tree cannot accurately
represent the meaning of consents and scopes in Auth, we have agreed
to pull scope parsing from the upcoming SDK release. The code is
remaining in-tree, but no longer provided via documented modules and
import paths. The changelog for it has been pulled, replaced by a more
modest set of changes to MutableScope, and the docs changes have been
further modified to remove any references to scope parsing.

Tests continue to exercise the scope tree parsing logic and all of the
utility of the Scope class as it was previously introduced.

* Apply suggestions from code review

Co-authored-by: Kurt McKee <contactme@kurtmckee.org>