Add support for new GCS collection doc versions (#675)
Support versions '1.6.0' and '1.7.0' via `guest_auth_policy_id` and
long `user_message` handling.

Internally, the `guest_auth_policy_id` is simply a
mapped-collection-only field matching the current structure, but the
`user_message` treatment requires a new structure.

DocumentWithInducedDatatype now defines `DATATYPE_VERSION_CALLBACKS`
as a secondary versioning mechanism alongside the
`DATATYPE_VERSION_IMPLICATIONS` mapping. The callbacks take the entire
object and return an optional version number tuple. This is an
extremely generic way of allowing a callback to apply arbitrary
criteria for the version numbering. In service of that change, the
protocol now requires `__getitem__` (trivially satisfied by all instances)
so that the contents can be checked.
The callbacks are defined as a classvar tuple.