Minor refinements to response classes (#680)
IterableResponse and ArrayResponse were not available as top-level
names exported by `globus_sdk`, whereas `GlobusHTTPResponse` was
available. This doesn't stem from any design decision, but is merely
an artifact of whatever evolved over time. It's potentially useful to
be able to pull up the response classes more directly -- at the very
least, it can make type annotations simpler and more accurate, with
little effort.

At the same time as this improvement is made, both ArrayResponse and
IterableResponse are given an improvement to the way that malformed
data is handled. Previously, payload data is `typing.cast` to the
correct type, ignoring potential runtime failures from trying to index
into or take the length of non-dict or non-list data. Instead,
explicitly check the types and raise TypeError if they are not
correct, as is already done by `ArrayResponse.__len__`. New tests
exercise the behavior.