Make `MutableScope.scope_string` public (#687)
This is a private attribute, but it turns out that there is at least
one legitimate use-case for creating MutableScope objects and then
accessing the `scope_string`:

In the CLI, we have a case in which we're tracking our scope
requirements as strings. If we wanted to add scope dependency info to
that mapping, it would be nice to do so as MutableScope objects.
However, on startup, we want to compare the top-level scopes against
the tokens we have in storage to ensure they are valid. This is most
naturally phrased as accessing the `scope_string` member of any
complex scopes we have in our requirements data.

We don't need to advertise the change heavily, but move this out of
an underscore-prefixed member and include it in the changelog.