Catch empty and unspecified `requested_scopes` (#690)
* Catch empty and unspecified `requested_scopes`

Primarily, this changes a runtime behavior today to catch and error on
an otherwise valid(ish) but undesirable condition. An empty scope string
or collection of scopes would previously be replaced with the SDK's
default scopes due to a loose "falsey" check. This is potentially
confusing or even wrong. A user trying intentionally to pass `""` will
find that their value is discarded in favor of the defaults. The new
behavior is that any falsey non-None value, (e.g. `[]`) will be
rejected with a usage error.

This is implemented in a new helper for handling the None check and
default behavior. In that helper, we also now have a deprecation
warning behavior inspired by the `sqlalchemy` 2.0 warnings and our
generally positive experience integrating against and
enabling/disabling those warnings.

The deprecation warning helpers define a warning class which inherits
from `Warning` rather than `DeprecationWarning` so that it won't be
suppressed by the default python warning filters. The warning is meant
to be emitted via a helper which checks an environment variable to see
if the "GLOBUS_SDK_V4_WARNINGS" has been set (enabling our v4
warnings). As a result, users won't see new warnings with current
usage or while following our examples and tutorial, but they will be
able to trivially enable the warnings by setting the env var.

New segments in the docs refer to the new env var and help users
opt-in. Tests cover the warning behavior and explore the helper
method a bit.

* Simplify v4 deprecation warnings and fix tests

Converting the deprecation warning behavior to always-on using
DeprecationWarning as the base class caused some tests (which trigger
the relevant warning) to fail. In these cases, exercise both runtime
behaviors with and without triggering the warning, but mark the
warning variant with a `pytest.warns` assertion.

In order to support some of the fixed tests, a new fixture is added in
`globus_sdk._testing`.