Add sphinx directive to list _testing data
This adds a directive to our internal sphinx extension to list the
data visible in `_testing`.

It uses the same helper as `AutoMethodList`, but
`EnumerateTestingFixtures` checks those fixtures against the
`_testing` registry and emits ReST for those which have defined test
cases.

The listing is done on a new method_list page, which follows the
existing norms and forms of the ("alpha") testing docs.