More testsuite cleanup: work on moving everything to _testing (#693)
* Remove unneeded things from tests.common

This module contained some constants which are no longer used in tests
and a helper which would be better provided as a pytest fixture.
Remove the constants and make the helper into a function fixture.

* Cleanup Local GCP tests and add data to fixtures

GlobusConnectPersonalOwnerInfo tests relied on the presence of
response payloads from get_identities with a few characteristics not
otherwise seen in the response data. One is a response for a
globusid.org user, and the other is an empty response. These two are
added to the test fixtures, and tests are converted to their use.

* Update base client tests to use _testing

And add `RegisteredResponse.replace()` as a method which allows use of
`responses.replace`. This is needed for some of the base client
tests.