Update TransferData.add_item to omit recursive (#696)
* Update TransferData.add_item to omit recursive

By default today, `add_item` will create a subdocument with
`{"recursive": false}`. This update defaults to omission of the
`"recursive"` key to match the new behavior being released by the
Transfer API.

This change is technically backwards incompatible in that we're
changing the document we generate. However, our previous default
(`False`) was chosen to match API behaviors and the new default
(omission) is chosen to match new API behaviors. This can impact
users, but is likely acceptable without a major version bump. To have
a noticeable negative impact, a user would have to not be specifying
`recursive` but also explicitly want the behavior of `recursive=False`.

The changelog entry encourages users to be explicit about their
desired usage if they are concerned.

* Weaken changelog note against setting 'recursive'

This tones down the statement in the changelog to better allow for the
possibility that setting the `recursive` flag explicitly can benefit
certain use-cases.

* Minor type annotation fix